Create server.js
Added the server.js file. Implemented a series of checks and validations methods for testing the entered URLs. From the original URL, after it passes the tests, I extract the domain name, original_url and generate a basic numeric based short_url for every entered URL. The first check to see if the entered URL points to a existing site, is to apply a dns.lookup(hostname, callback). After it passes the first important check, next test is to see if the entered URL uses a valid protocol, using a match with a specific regExp. To save the documents in the mongoDB, I'm using the .create method that get's all docs from the savedSites array, that's getting entries from every passing URL post. Next step is to implement a method to check each new entry against the ones already in the DB, to avoid duplicates. Also important is to get the last entry in the DB to find the last number used as a short_url to get the count variable for a new entry.