Update server.js
Coded and implemented the check for duplicates method and logic, the getLastCount() function that retrieves the short_url value for the last entry in the DB and updates the count variable on page load and finally completed the Route method for redirecting the page based on the user's parameter input. The project is now final and completely functional. The project is ready to be submitted to FCC for review.